remove SchemaAnnotationHandlerClassUtil and SchemaAnnotationHandlerUtil since they are not used
