+++ restli-tools/src/main/java/com/linkedin/restli/tools/annotation/SchemaAnnotationValidatorCmdLineApp.java
@@ -25,9 +25,14 @@ import com.linkedin.pegasus.generator.DataSchemaParser;
 import com.linkedin.restli.internal.tools.RestLiToolsUtils;
 import java.io.File;
 import java.io.IOException;
+import java.net.URL;
+import java.net.URLClassLoader;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.StringTokenizer;
 import java.util.stream.Collectors;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -111,7 +116,35 @@ public class SchemaAnnotationValidatorCmdLineApp
       System.exit(1);
     }
 
-    List<SchemaAnnotationHandler> handlers = SchemaAnnotationHandlerUtil.getSchemaAnnotationHandlers(handlerJarPaths, handlerClassNames);
+    List<String> handlerJarPathsArray = parsePaths(handlerJarPaths);
+    // Use Jar Paths to initiate URL class loaders.
+    ClassLoader classLoader = new URLClassLoader(handlerJarPathsArray.stream().map(str -> {
+      try
+      {
+        return Paths.get(str).toUri().toURL();
+      } catch (Exception e)
+      {
+        _log.error("URL {} parsing failed", str, e);
+      }
+      return null;
+    }).filter(Objects::nonNull).toArray(URL[]::new));
+    List<SchemaAnnotationHandler> handlers = new ArrayList<>();
+    for (String className: parsePaths(handlerClassNames))
+    {
+      try
+      {
+        Class<?> handlerClass = Class.forName(className, false, classLoader);
+        SchemaAnnotationHandler handler = (SchemaAnnotationHandler) handlerClass.newInstance();
+        handlers.add(handler);
+        _log.info("added handler {} for annotation namespace \"{}\"", className, handler.getAnnotationNamespace());
+      }
+      catch (Exception e)
+      {
+        _log.error("Error instantiating handler class {} ", className, e);
+        // fail even just one handler fails
+        throw new IllegalStateException("ValidateSchemaAnnotation task failed");
+      }
+    }
 
     boolean hasError = false;
     List<String> schemaWithFailures = new ArrayList<>();
@@ -141,6 +174,20 @@ public class SchemaAnnotationValidatorCmdLineApp
       // Throw exception at the end if any of
       throw new IllegalStateException("ValidateSchemaAnnotation task failed");
     }
+  }
+
+  private static List<String> parsePaths(String pathAsStr)
+  {
+    List<String> list = new ArrayList<>();
+    if (pathAsStr != null)
+    {
+      StringTokenizer tokenizer = new StringTokenizer(pathAsStr, DEFAULT_PATH_SEPARATOR);
+      while (tokenizer.hasMoreTokens())
+      {
+        list.add(tokenizer.nextToken());
+      }
+    }
+    return list;
   }
 
   private static List<DataSchema> parseSchemas(String resolverPath, String modelsLocation) throws IOException
